halfway through refactoring

+++ Lib/DataTypes/MonteCarlo/McPerson.cs
@@ -10,10 +10,10 @@ namespace Lib.DataTypes.MonteCarlo
         public  Guid Id { get; set; }
         public  string Name { get; set; }
         public  LocalDateTime BirthDate {  get; set; }
-        public  decimal AnnualSalary { get; set; }
-        public  decimal AnnualBonus { get; set; }
-        public  decimal MonthlyFullSocialSecurityBenefit { get; set; }
-        public decimal Annual401kMatchPercent { get; set; }
+        public  long AnnualSalary { get; set; }
+        public  long AnnualBonus { get; set; }
+        public  long MonthlyFullSocialSecurityBenefit { get; set; }
+        public long Annual401kMatchPercent { get; set; }
         public List<McInvestmentAccount> InvestmentAccounts { get; set; } = [];
         public List<McDebtAccount> DebtAccounts { get; set; } = [];
         public List<McModel> Models { get; set; } = [];
